Module Storage.Liquidity_baking

module Escape_ema : Tezos_raw_protocol_alpha.Storage_sigs.Single_data_storage with type t := Tezos_raw_protocol_alpha.Raw_context.t and type value = int32

Exponential moving average (ema) of flags set in protocol_data.contents. If at any block it's above the threshold set in constants, liquidity baking permanently shuts off. *

module Cpmm_address : Tezos_raw_protocol_alpha.Storage_sigs.Single_data_storage with type t := Tezos_raw_protocol_alpha.Raw_context.t and type value = Tezos_raw_protocol_alpha.Contract_repr.t

Constant product market maker contract that receives liquidity baking subsidy. *