Module Tezos_raw_protocol_alpha

module Alpha_context : sig ... end

module Alpha_services : sig ... end

module Amendment : sig ... end

Only delegates with at least one roll take part in the amendment procedure. It works as follows:

module Apply : sig ... end

module Apply_results : sig ... end

Result of applying an operation, can be used for experimenting with protocol updates, by clients to print out a summary of the operation at pre-injection simulation and at confirmation time, and by block explorers.

module Baking : sig ... end

module Blinded_public_key_hash : sig ... end

module Block_header_repr : sig ... end

module Bootstrap_storage : sig ... end

module Commitment_repr : sig ... end

module Commitment_storage : sig ... end

module Constants_repr : sig ... end

module Constants_services : sig ... end

module Constants_storage : sig ... end

module Contract_hash : sig ... end

module Contract_repr : sig ... end

module Contract_services : sig ... end

module Contract_storage : sig ... end

module Cycle_repr : sig ... end

module Delegate_services : sig ... end

module Delegate_storage : sig ... end

module Fees_storage : sig ... end

module Fitness_repr : sig ... end

module Fitness_storage : sig ... end

module Fixed_point_repr : sig ... end

module Gas_limit_repr : sig ... end

module Init_storage : sig ... end

module Lazy_storage_diff : sig ... end

module Lazy_storage_kind : sig ... end

module Level_repr : sig ... end

module Level_storage : sig ... end

module Liquidity_baking_cpmm : sig ... end

module Liquidity_baking_lqt : sig ... end

module Liquidity_baking_migration : sig ... end

module Liquidity_baking_repr : sig ... end

module Main : sig ... end

module Manager_repr : sig ... end

module Michelson_v1_gas : sig ... end

module Michelson_v1_primitives : sig ... end

module Migration_repr : sig ... end

module Misc : sig ... end

module Nonce_hash : sig ... end

module Nonce_storage : sig ... end

module Operation_repr : sig ... end

module Parameters_repr : sig ... end

module Period_repr : sig ... end

module Raw_context : sig ... end

module Raw_context_intf : sig ... end

module Raw_level_repr : sig ... end

module Receipt_repr : sig ... end

module Roll_repr : sig ... end

module Roll_storage : sig ... end

module Sapling_repr : sig ... end

module Sapling_services : sig ... end

module Sapling_storage : sig ... end

module Sapling_validator : sig ... end

module Saturation_repr : sig ... end

module Script_expr_hash : sig ... end

module Script_int_repr : sig ... end

module Script_interpreter : sig ... end

This is the Michelson interpreter.

module Script_interpreter_defs : sig ... end

module Script_ir_annot : sig ... end

module Script_ir_translator : sig ... end

module Script_repr : sig ... end

module Script_tc_errors : sig ... end

module Script_tc_errors_registration : sig ... end

module Script_timestamp_repr : sig ... end

module Script_typed_ir : sig ... end

module Seed_repr : sig ... end

module Seed_storage : sig ... end

module Services_registration : sig ... end

module State_hash : sig ... end

module Storage : sig ... end

module Storage_costs : sig ... end

module Storage_description : sig ... end

module Storage_functors : sig ... end

Tezos Protocol Implementation - Typed storage builders.

module Storage_sigs : sig ... end

module Tez_repr : sig ... end

module Time_repr : sig ... end

module Vote_repr : sig ... end

module Vote_storage : sig ... end

module Voting_period_repr : sig ... end

module Voting_period_storage : sig ... end

module Voting_services : sig ... end