Module Tezos_raw_protocol_alpha.Alpha_services

This declares Protocol RPC services.

Protocol RPC services are read-only, and support querying the state of the ledger (including information such as existing contracts, delegation, voting, and so on), at a given block height.

This is a mostly internal module used from rpc_services in Main.

module Seed_computation : sig ... end

module Seed : sig ... end

module Nonce : sig ... end

module Constants = Constants_services

module Voting = Voting_services

module Sapling = Sapling_services

module Liquidity_baking : sig ... end

module Cache : sig ... end

module Denunciations : sig ... end

val register : unit -> unit