Module Tezos_raw_protocol_alpha.Apply_results

Types representing results of applying an operation.

These are used internally by Apply, and can be used for experimenting with protocol updates, by clients to print out a summary of the operation at pre-injection simulation and at confirmation time, and by block explorers.

type 'kind operation_metadata = {

contents : 'kind contents_result_list;

}

Result of applying a Operation.t. Follows the same structure.

and packed_operation_metadata =

| Operation_metadata : 'kind operation_metadata -> packed_operation_metadata

| No_operation_metadata : packed_operation_metadata

and 'kind contents_result_list =

| Single_result : 'kind contents_result -> 'kind contents_result_list

| Cons_result : 'kind Alpha_context.Kind.manager contents_result * 'rest Alpha_context.Kind.manager contents_result_list -> ('kind * 'rest) Alpha_context.Kind.manager contents_result_list

Result of applying a Operation.contents_list. Follows the same structure.

and packed_contents_result_list =

| Contents_result_list : 'kind contents_result_list -> packed_contents_result_list

and 'kind contents_result =

| Preattestation_result : {

balance_updates : Alpha_context.Receipt.balance_updates;

delegate : Tezos_protocol_environment_alpha.Signature.public_key_hash;

consensus_key : Tezos_protocol_environment_alpha.Signature.public_key_hash;

consensus_power : int;

} -> Alpha_context.Kind.preattestation contents_result

| Attestation_result : {

balance_updates : Alpha_context.Receipt.balance_updates;

delegate : Tezos_protocol_environment_alpha.Signature.public_key_hash;

consensus_key : Tezos_protocol_environment_alpha.Signature.public_key_hash;

consensus_power : int;

} -> Alpha_context.Kind.attestation contents_result

| Seed_nonce_revelation_result : Alpha_context.Receipt.balance_updates -> Alpha_context.Kind.seed_nonce_revelation contents_result

| Vdf_revelation_result : Alpha_context.Receipt.balance_updates -> Alpha_context.Kind.vdf_revelation contents_result

| Double_attestation_evidence_result : {

forbidden_delegate : Tezos_protocol_environment_alpha.Signature.public_key_hash option;

balance_updates : Alpha_context.Receipt.balance_updates;

} -> Alpha_context.Kind.double_attestation_evidence contents_result

| Double_preattestation_evidence_result : {

forbidden_delegate : Tezos_protocol_environment_alpha.Signature.public_key_hash option;

balance_updates : Alpha_context.Receipt.balance_updates;

} -> Alpha_context.Kind.double_preattestation_evidence contents_result

| Double_baking_evidence_result : {

forbidden_delegate : Tezos_protocol_environment_alpha.Signature.public_key_hash option;

balance_updates : Alpha_context.Receipt.balance_updates;

} -> Alpha_context.Kind.double_baking_evidence contents_result

| Dal_entrapment_evidence_result : {

balance_updates : Alpha_context.Receipt.balance_updates;

} -> Alpha_context.Kind.dal_entrapment_evidence contents_result

| Activate_account_result : Alpha_context.Receipt.balance_updates -> Alpha_context.Kind.activate_account contents_result

| Proposals_result : Alpha_context.Kind.proposals contents_result

| Ballot_result : Alpha_context.Kind.ballot contents_result

| Drain_delegate_result : {

balance_updates : Alpha_context.Receipt.balance_updates;

allocated_destination_contract : bool;

} -> Alpha_context.Kind.drain_delegate contents_result

| Manager_operation_result : {

balance_updates : Alpha_context.Receipt.balance_updates;

operation_result : 'kind manager_operation_result;

internal_operation_results : Apply_internal_results.packed_internal_operation_result list;

} -> 'kind Alpha_context.Kind.manager contents_result

Result of applying an Operation.contents. Follows the same structure.

and packed_contents_result =

| Contents_result : 'kind contents_result -> packed_contents_result

and 'kind manager_operation_result = ('kind, 'kind Alpha_context.Kind.manager, 'kind successful_manager_operation_result) Apply_operation_result.operation_result

and successful_transaction_result = Apply_internal_results.successful_transaction_result

Result of applying a transaction.

and successful_origination_result = Apply_internal_results.successful_origination_result

Result of applying an origination.

and _ successful_manager_operation_result =

| Reveal_result : {

consumed_gas : Alpha_context.Gas.Arith.fp;

} -> Alpha_context.Kind.reveal successful_manager_operation_result

| Transaction_result : successful_transaction_result -> Alpha_context.Kind.transaction successful_manager_operation_result

| Origination_result : successful_origination_result -> Alpha_context.Kind.origination successful_manager_operation_result

| Delegation_result : {

consumed_gas : Alpha_context.Gas.Arith.fp;

balance_updates : Alpha_context.Receipt.balance_updates;

} -> Alpha_context.Kind.delegation successful_manager_operation_result

| Register_global_constant_result : {

balance_updates : Alpha_context.Receipt.balance_updates;

consumed_gas : Alpha_context.Gas.Arith.fp;

size_of_constant : Tezos_protocol_environment_alpha.Z.t;

global_address : Script_expr_hash.t;

} -> Alpha_context.Kind.register_global_constant successful_manager_operation_result

| Set_deposits_limit_result : {

consumed_gas : Alpha_context.Gas.Arith.fp;

} -> Alpha_context.Kind.set_deposits_limit successful_manager_operation_result

| Increase_paid_storage_result : {

balance_updates : Alpha_context.Receipt.balance_updates;

consumed_gas : Alpha_context.Gas.Arith.fp;

} -> Alpha_context.Kind.increase_paid_storage successful_manager_operation_result

| Update_consensus_key_result : {

consumed_gas : Alpha_context.Gas.Arith.fp;

} -> Alpha_context.Kind.update_consensus_key successful_manager_operation_result

| Transfer_ticket_result : {

balance_updates : Alpha_context.Receipt.balance_updates;

ticket_receipt : Ticket_receipt.t;

consumed_gas : Alpha_context.Gas.Arith.fp;

paid_storage_size_diff : Tezos_protocol_environment_alpha.Z.t;

} -> Alpha_context.Kind.transfer_ticket successful_manager_operation_result

| Dal_publish_commitment_result : {

slot_header : Alpha_context.Dal.Slot.Header.t;

consumed_gas : Alpha_context.Gas.Arith.fp;

} -> Alpha_context.Kind.dal_publish_commitment successful_manager_operation_result

| Sc_rollup_originate_result : {

balance_updates : Alpha_context.Receipt.balance_updates;

address : Alpha_context.Sc_rollup.Address.t;

genesis_commitment_hash : Alpha_context.Sc_rollup.Commitment.Hash.t;

consumed_gas : Alpha_context.Gas.Arith.fp;

size : Tezos_protocol_environment_alpha.Z.t;

} -> Alpha_context.Kind.sc_rollup_originate successful_manager_operation_result

| Sc_rollup_add_messages_result : {

consumed_gas : Alpha_context.Gas.Arith.fp;

} -> Alpha_context.Kind.sc_rollup_add_messages successful_manager_operation_result

| Sc_rollup_cement_result : {

consumed_gas : Alpha_context.Gas.Arith.fp;

inbox_level : Alpha_context.Raw_level.t;

commitment_hash : Alpha_context.Sc_rollup.Commitment.Hash.t;

} -> Alpha_context.Kind.sc_rollup_cement successful_manager_operation_result

| Sc_rollup_publish_result : {

consumed_gas : Alpha_context.Gas.Arith.fp;

staked_hash : Alpha_context.Sc_rollup.Commitment.Hash.t;

published_at_level : Alpha_context.Raw_level.t;

balance_updates : Alpha_context.Receipt.balance_updates;

} -> Alpha_context.Kind.sc_rollup_publish successful_manager_operation_result

| Sc_rollup_refute_result : {

consumed_gas : Alpha_context.Gas.Arith.fp;

game_status : Alpha_context.Sc_rollup.Game.status;

balance_updates : Alpha_context.Receipt.balance_updates;

} -> Alpha_context.Kind.sc_rollup_refute successful_manager_operation_result

| Sc_rollup_timeout_result : {

consumed_gas : Alpha_context.Gas.Arith.fp;

game_status : Alpha_context.Sc_rollup.Game.status;

balance_updates : Alpha_context.Receipt.balance_updates;

} -> Alpha_context.Kind.sc_rollup_timeout successful_manager_operation_result

| Sc_rollup_execute_outbox_message_result : {

balance_updates : Alpha_context.Receipt.balance_updates;

ticket_receipt : Ticket_receipt.t;

whitelist_update : Alpha_context.Sc_rollup.Whitelist.update option;

consumed_gas : Alpha_context.Gas.Arith.fp;

paid_storage_size_diff : Tezos_protocol_environment_alpha.Z.t;

} -> Alpha_context.Kind.sc_rollup_execute_outbox_message successful_manager_operation_result

| Sc_rollup_recover_bond_result : {

balance_updates : Alpha_context.Receipt.balance_updates;

consumed_gas : Alpha_context.Gas.Arith.fp;

} -> Alpha_context.Kind.sc_rollup_recover_bond successful_manager_operation_result

| Zk_rollup_origination_result : {

balance_updates : Alpha_context.Receipt.balance_updates;

originated_zk_rollup : Alpha_context.Zk_rollup.t;

consumed_gas : Alpha_context.Gas.Arith.fp;

storage_size : Tezos_protocol_environment_alpha.Z.t;

} -> Alpha_context.Kind.zk_rollup_origination successful_manager_operation_result

| Zk_rollup_publish_result : {

balance_updates : Alpha_context.Receipt.balance_updates;

consumed_gas : Alpha_context.Gas.Arith.fp;

paid_storage_size_diff : Tezos_protocol_environment_alpha.Z.t;

} -> Alpha_context.Kind.zk_rollup_publish successful_manager_operation_result

| Zk_rollup_update_result : {

balance_updates : Alpha_context.Receipt.balance_updates;

consumed_gas : Alpha_context.Gas.Arith.fp;

paid_storage_size_diff : Tezos_protocol_environment_alpha.Z.t;

} -> Alpha_context.Kind.zk_rollup_update successful_manager_operation_result

Result of applying an external manager_operation_content.

and packed_successful_manager_operation_result =

| Successful_manager_result : 'kind successful_manager_operation_result -> packed_successful_manager_operation_result

val pack_migration_operation_results : Alpha_context.Migration.origination_result list -> packed_successful_manager_operation_result list

val operation_metadata_encoding : packed_operation_metadata Tezos_protocol_environment_alpha.Data_encoding.t

Serializer for packed_operation_result.

val operation_data_and_metadata_encoding : (Alpha_context.Operation.packed_protocol_data * packed_operation_metadata) Tezos_protocol_environment_alpha.Data_encoding.t

type 'kind contents_and_result_list =

| Single_and_result : 'kind Alpha_context.contents * 'kind contents_result -> 'kind contents_and_result_list

| Cons_and_result : 'kind Alpha_context.Kind.manager Alpha_context.contents * 'kind Alpha_context.Kind.manager contents_result * 'rest Alpha_context.Kind.manager contents_and_result_list -> ('kind * 'rest) Alpha_context.Kind.manager contents_and_result_list

type packed_contents_and_result_list =

| Contents_and_result_list : 'kind contents_and_result_list -> packed_contents_and_result_list

val contents_and_result_list_encoding : packed_contents_and_result_list Tezos_protocol_environment_alpha.Data_encoding.t

val pack_contents_list : 'kind Alpha_context.contents_list -> 'kind contents_result_list -> 'kind contents_and_result_list

val unpack_contents_list : 'kind contents_and_result_list -> 'kind Alpha_context.contents_list * 'kind contents_result_list

val to_list : packed_contents_result_list -> packed_contents_result list

type ('a, 'b) eq =

| Eq : ('a, 'a) eq

val kind_equal_list : 'kind Alpha_context.contents_list -> 'kind2 contents_result_list -> ('kind, 'kind2) eq option

type block_metadata = {

proposer : Alpha_context.Consensus_key.t;

baker : Alpha_context.Consensus_key.t;

level_info : Alpha_context.Level.t;

voting_period_info : Alpha_context.Voting_period.info;

nonce_hash : Nonce_hash.t option;

consumed_gas : Alpha_context.Gas.Arith.fp;

deactivated : Tezos_protocol_environment_alpha.Signature.Public_key_hash.t list;

balance_updates : Alpha_context.Receipt.balance_updates;

liquidity_baking_toggle_ema : Alpha_context.Per_block_votes.Liquidity_baking_toggle_EMA.t;

adaptive_issuance_vote_ema : Alpha_context.Per_block_votes.Adaptive_issuance_launch_EMA.t;

adaptive_issuance_launch_cycle : Alpha_context.Cycle.t option;

implicit_operations_results : packed_successful_manager_operation_result list;

dal_attestation : Alpha_context.Dal.Attestation.t;

}

val block_metadata_encoding : block_metadata Tezos_protocol_environment_alpha.Data_encoding.encoding