Module Tezos_raw_protocol_alpha.Baking

type Tezos_protocol_environment_alpha.Error_monad.error +=

| Insufficient_attestation_power of {

attestation_power : int;

consensus_threshold : int;

}

type ordered_slots = private {

delegate : Tezos_protocol_environment_alpha.Signature.public_key_hash;

consensus_key : Tezos_protocol_environment_alpha.Signature.public_key_hash;

slots : Alpha_context.Slot.t list;

}

val attesting_rights : Alpha_context.context -> Alpha_context.Level.t -> (Alpha_context.context * ordered_slots Tezos_protocol_environment_alpha.Signature.Public_key_hash.Map.t) Tezos_protocol_environment_alpha.Error_monad.tzresult Tezos_protocol_environment_alpha.Lwt.t

For a given level computes who has the right to include an attestation in the next block.

returns
map from delegates with such rights to their attesting slots, in increasing order.

This function is only used by the 'validators' RPC.

val attesting_rights_by_first_slot : Alpha_context.context -> Alpha_context.Level.t -> (Alpha_context.context * (Alpha_context.Consensus_key.pk * int * int) Alpha_context.Slot.Map.t) Tezos_protocol_environment_alpha.Error_monad.tzresult Tezos_protocol_environment_alpha.Lwt.t

Computes attesting rights for a given level.

returns
map from allocated first slots to their owner's public key, consensus attesting power, and DAL attesting power.

val bonus_baking_reward : Alpha_context.context -> attestation_power:int -> Alpha_context.Tez.t Tezos_protocol_environment_alpha.Error_monad.tzresult

Computes the bonus baking reward depending on the attestation power.