Module Make.Internal_for_tests

val empty : capacity:int64 -> next_index:int64 -> t

A more flexible empty function for testing purpose.

val keys : t -> key list

keys t returns the keys of the entries stored in t in the order of their insertion.