Module Dal_slot_repr.Header

The module below provides some data types and helper functions for DAL commitments published at some level on some slot index. There are mainly three important levels for a DAL commitment successfully included in a block:

`published_level`: The level of the block that contains the DAL publish commitment operation. Such operations are typically injected on top of the block whose level is `published_level - 1` or earlier.

`attested_level`: The level of the block / context that includes the attestation status of a slot published `attestation_lag` levels earlier. Formally, `attested_level = published_level + attestation_lag`.

`attestation_level`: The level of the block on top of which DAL attestations are injected for a commitment published at published_level. Formally, attestation_level = attested_level - 1 = published_level + attestation_lag - 1.

type id = {

published_level : Raw_level_repr.t;

index : Dal_slot_index_repr.t;

}

For Layer-1, a slot is identified by the level at which it is published and the slot's index.

type t = {

id : id;

commitment : Commitment.t;

}

For Layer-1, a slot is described by its slot id and the slot's commitment.

val id_encoding : id Tezos_protocol_environment_alpha.Data_encoding.t

encoding for values of type id.

val encoding : t Tezos_protocol_environment_alpha.Data_encoding.t

encoding for values of type t.

val pp_id : Tezos_protocol_environment_alpha.Format.formatter -> id -> unit

pretty-printer for values of type id.

val pp : Tezos_protocol_environment_alpha.Format.formatter -> t -> unit

pretty-printer for values of type t.

val equal : t -> t -> bool

equal function for values of type t.

val slot_id_equal : id -> id -> bool

equal function for values of type id.

val verify_commitment : Tezos_protocol_environment_alpha.Dal.t -> Commitment.t -> Commitment_proof.t -> bool Tezos_protocol_environment_alpha.Error_monad.tzresult

verify_commitment cryptobox commitment commitment_proof check that for the given commitment, the commitment proof is correct using the cryptbox primitives.