Module Dal_slot_repr.Shard_with_proof

type t = {

shard : Tezos_protocol_environment_alpha.Dal.shard;

proof : Tezos_protocol_environment_alpha.Dal.shard_proof;

}

type Tezos_protocol_environment_alpha.Error_monad.error +=

| Dal_shard_proof_error of string

val encoding : t Tezos_protocol_environment_alpha.Data_encoding.t

val verify : Tezos_protocol_environment_alpha.Dal.t -> Commitment.t -> t -> unit Tezos_protocol_environment_alpha.Error_monad.tzresult

val share_is_trap : Tezos_protocol_environment_alpha.Signature.Public_key_hash.t -> Tezos_protocol_environment_alpha.Dal.share -> traps_fraction:Tezos_protocol_environment_alpha.Q.t -> bool Tezos_protocol_environment_alpha.Error_monad.tzresult

share_is_trap pkh share ~traps_fraction determines whether the given share is classified as a "trap" for the delegate identified by pkh, based on the fraction traps_fraction of shards that should be traps.

The function computes the hash of the concatenation of pkh and share, denoted as `hash(pkh . share)`, where the dot represents concatenation. It then checks if this hash value is less than `traps_fraction * 2^n`, where `n` is the (fixed) bit size of the hash.

The function returns:

Ok true if the share is considered a trap for the given pkh.

Ok false if the share is not a trap.

Error Dal_encoding_error_in_share_is_trap if there is an issue encoding pkh or share.

This function assumes traps_fraction is valid (i.e., a rational number within 0, 1).