Module Tezos_raw_protocol_alpha.Delegate_missed_attestations_storage

This modules deals with delegates' participation in consensus.

This module is responsible for maintaining the Storage.Contract.Missed_attestations table.

val expected_slots_for_given_active_stake : Raw_context.t -> total_active_stake_weight:int64 -> active_stake_weight:int64 -> int

Computes the number of validator slots that a delegate is expected to be allocated during a cycle. This number is proportional to its active stake wrt to total active stake.

val expected_dal_shards_for_given_active_stake : Raw_context.t -> total_active_stake_weight:int64 -> active_stake_weight:int64 -> int

Computes the number of DAL shards that a delegate is expected to be allocated during a cycle. This number is proportional to its active stake wrt to total active stake.

val remove_total_dal_attested_slots : Raw_context.t -> Raw_context.t Tezos_protocol_environment_alpha.Lwt.t

Removes the Storage.Dal.Total_attested_slots value from the context.

type level_participation =

| Participated

| Didn't_participate

val record_attesting_participation : Raw_context.t -> delegate:Tezos_protocol_environment_alpha.Signature.Public_key_hash.t -> participation:level_participation -> attesting_power:int -> Raw_context.t Tezos_protocol_environment_alpha.Error_monad.tzresult Tezos_protocol_environment_alpha.Lwt.t

Update the participation of a delegate as a validator in the current cycle with its participation at the current level.

val record_dal_participation : Raw_context.t -> delegate:Tezos_protocol_environment_alpha.Signature.Public_key_hash.t -> number_of_attested_slots:int -> Raw_context.t Tezos_protocol_environment_alpha.Error_monad.tzresult Tezos_protocol_environment_alpha.Lwt.t

Update the participation of a delegate as a DAL attester in the current cycle with its participation (ie the number of DAL slots it attested) at the current level.

val record_baking_activity_and_pay_rewards_and_fees : Raw_context.t -> payload_producer:Tezos_protocol_environment_alpha.Signature.Public_key_hash.t -> block_producer:Tezos_protocol_environment_alpha.Signature.Public_key_hash.t -> baking_reward:Tez_repr.t -> reward_bonus:Tez_repr.t option -> (Raw_context.t * Receipt_repr.balance_updates) Tezos_protocol_environment_alpha.Error_monad.tzresult Tezos_protocol_environment_alpha.Lwt.t

Sets the payload and block producer as active. Pays the baking reward and the fees to the payload producer and the reward bonus to the payload producer (if the reward_bonus is not None).

val check_and_reset_delegate_participation : Raw_context.t -> Tezos_protocol_environment_alpha.Signature.Public_key_hash.t -> (Raw_context.t * bool) Tezos_protocol_environment_alpha.Error_monad.tzresult Tezos_protocol_environment_alpha.Lwt.t

Check that a delegate participated enough in the last cycle (returns true if it did), and then reset the participation for preparing the next cycle.

val get_and_reset_delegate_dal_participation : Raw_context.t -> Tezos_protocol_environment_alpha.Signature.Public_key_hash.t -> (Raw_context.t * int32) Tezos_protocol_environment_alpha.Error_monad.tzresult Tezos_protocol_environment_alpha.Lwt.t

Retrieve the number of DAL slots a delegate DAL-attested during the last cycle, and then reset the participation for preparing the next cycle.

module For_RPC : sig ... end