Module Delegate_slashed_deposits_storage.For_RPC

val get_estimated_shared_pending_slashed_amount : Raw_context.t -> Tezos_protocol_environment_alpha.Signature.public_key_hash -> Tez_repr.t Tezos_protocol_environment_alpha.Error_monad.tzresult Tezos_protocol_environment_alpha.Lwt.t

get_estimated_shared_pending_slashed_amount ctxt delegate returns the estimated shared pending slashed amount of the given delegate according to the currently available denunciations.

val get_estimated_own_pending_slashed_amount : Raw_context.t -> Contract_repr.t -> Tez_repr.t Tezos_protocol_environment_alpha.Error_monad.tzresult Tezos_protocol_environment_alpha.Lwt.t

get_estimated_own_pending_slashed_amount ctxt contract returns the estimated own pending slashed amount of the given contract according to the currently available denunciations.