Module Tezos_raw_protocol_alpha.Frozen_staker_repr

type t = private

| Baker of Tezos_protocol_environment_alpha.Signature.public_key_hash
(*
The baker itself.
*)

| Single_staker of {

staker : Contract_repr.t;

delegate : Tezos_protocol_environment_alpha.Signature.public_key_hash;

}
(*
A single staker, cannot be the delegate.
*)

| Shared_between_stakers of {

delegate : Tezos_protocol_environment_alpha.Signature.public_key_hash;

}
(*
The delegate's stakers simultaneously (delegate excluded).
*)

| Baker_edge of Tezos_protocol_environment_alpha.Signature.public_key_hash
(*
The baker edge over its stakers rewards.
*)

Adding and removing stake can be done from/toward the funds

a delegate directly stakes,

one of its stakers stakes,

all its stakers stakes at once,

or from the edge a delegate may have on its stakers rewards. We need to distinguish these cases to enforce the staking over baking limit.

val baker : Tezos_protocol_environment_alpha.Signature.public_key_hash -> t

val baker_edge : Tezos_protocol_environment_alpha.Signature.public_key_hash -> t

val single_staker : staker:Contract_repr.t -> delegate:Tezos_protocol_environment_alpha.Signature.public_key_hash -> t

val shared_between_stakers : delegate:Tezos_protocol_environment_alpha.Signature.public_key_hash -> t

val encoding : t Tezos_protocol_environment_alpha.Data_encoding.t

val compare : t -> t -> int