Module Tezos_raw_protocol_alpha.Gas_input_size

Gas_input_size includes the definitions for the different sizes used in the gas models of the protocol. They do not always represent memory sizes, but rather they can be seen as an information size. They are tailored to the models that use them, and should not be used for anything other than gas computation.

include module type of Gas_comparable_input_size

type t = int

type micheline_size = {

traversal : t;

int_bytes : t;

string_bytes : t;

}

val encoding : t Tezos_protocol_environment_alpha.Data_encoding.encoding

val micheline_size_encoding : micheline_size Tezos_protocol_environment_alpha.Data_encoding.encoding

val zero : t

val add : t -> t -> t

val pp : Tezos_protocol_environment_alpha.Format.formatter -> t -> unit

val pp_micheline_size : Tezos_protocol_environment_alpha.Format.formatter -> micheline_size -> unit

val to_int : t -> int

val of_int : int -> t

val integer : 'a Script_int.num -> t

val string : string -> t

val script_string : Script_string.t -> t

val bytes : Tezos_protocol_environment_alpha.Bytes.t -> t

val mutez : Alpha_context.Tez.t -> t

val timestamp : Script_timestamp.t -> t

val size_of_comparable_value : 'a Script_typed_ir.comparable_ty -> 'a -> t

val list : 'a Script_list.t -> t

val set : 'a Script_typed_ir.set -> t

val map : ('a, 'b) Script_typed_ir.map -> t

val of_micheline : ('a, 'b) Tezos_protocol_environment_alpha.Micheline.node -> micheline_size

val sapling_transaction_inputs : Alpha_context.Sapling.transaction -> t

val sapling_transaction_outputs : Alpha_context.Sapling.transaction -> t

val sapling_transaction_bound_data : Alpha_context.Sapling.transaction -> t