Module Tezos_raw_protocol_alpha.Lazy_storage_diff

module Env = Tezos_protocol_environment_alpha.Environment

module Err = Env.Error_monad

See Lazy_storage_kind for an introduction on lazy storage.

This module defines operations on lazy storage types and diffs.

type ('id, 'alloc) init =

| Existing

| Copy of {

src : 'id;

}

| Alloc of 'alloc

type ('id, 'alloc, 'updates) diff =

| Remove

| Update of {

init : ('id, 'alloc) init;

updates : 'updates;

}

type diffs_item = private

| Item : ('i, 'a, 'u) Lazy_storage_kind.t * 'i * ('i, 'a, 'u) diff -> diffs_item

val make : ('i, 'a, 'u) Lazy_storage_kind.t -> 'i -> ('i, 'a, 'u) diff -> diffs_item

type diffs = diffs_item list

val encoding : diffs Env.Data_encoding.t

val apply : Raw_context.t -> diffs -> (Raw_context.t * Env.Z.t) Err.tzresult Env.Lwt.t

The returned Z.t is the size added by the application of the diffs.

val fresh : ('id, _, _) Lazy_storage_kind.t -> temporary:bool -> Raw_context.t -> (Raw_context.t * 'id) Err.tzresult Env.Lwt.t

val init : Raw_context.t -> Raw_context.t Err.tzresult Env.Lwt.t

Initializes the storage for all lazy storage kind. This is useful for genesis only. Protocol updates need to initialize new lazy storage kinds.

val cleanup_temporaries : Raw_context.t -> Raw_context.t Env.Lwt.t

val cleanup_edo_florence_dangling_lazy_storage : Raw_context.t -> Raw_context.t Env.Lwt.t