Module Tezos_raw_protocol_alpha.Manager_repr

module Env = Tezos_protocol_environment_alpha.Environment

type manager_key =

| Hash of Env.Signature.Public_key_hash.t

| Public_key of Env.Signature.Public_key.t

The public key of the manager of a contract is reveled only after the first operation. At Origination time, the manager provides only the hash of its public key that is stored in the contract. When the public key is actually revealed, the public key instead of the hash of the key

type t = manager_key

val encoding : t Env.Data_encoding.encoding