Module Sapling_storage.Roots

Bounded queue of roots. The full size is initialized with the default uncommitted root, that's why roots storage doesn't need to be carbonated. A maximum of one new root is added per protocol level. If multiple transactions for the same shielded pool are processed during the same contract call or several calls in the same block, only the last root will be stored. This property prevents transactions in the same block from depending on each other and guarantees that a transaction will be valid for a least two hours (hence the 120 size) after being forged.

val size : int32

val get : Raw_context.t -> Lazy_storage_kind.Sapling_state.Id.t -> (Storage.Sapling.Roots.value, Storage_sigs.Err.error Storage_sigs.Err.trace) Stdlib.result Lwt.t

val init : Raw_context.t -> Lazy_storage_kind.Sapling_state.Id.t -> (Raw_context.t, Storage_sigs.Err.error Storage_sigs.Err.trace) Stdlib.result Lwt.t

val mem : Raw_context.t -> Lazy_storage_kind.Sapling_state.Id.t -> Env.Sapling.Hash.t -> (bool, Storage_sigs.Err.error Storage_sigs.Err.trace) Stdlib.result Lwt.t

val add : Raw_context.t -> Lazy_storage_kind.Sapling_state.Id.t -> Storage.Sapling.Roots.value -> (Raw_context.t, Storage_sigs.Err.error Storage_sigs.Err.trace) Stdlib.result Lwt.t