Module Tezos_raw_protocol_alpha.Sc_rollup_arith

This module provides a temporary toy rollup to be used as a demo.

module type S = sig ... end

This rollup is a stack machine equipped with addition.

module Protocol_implementation : S with type context = Tezos_protocol_environment_alpha.Context.t and type state = Tezos_protocol_environment_alpha.Context.tree and type proof = Tezos_protocol_environment_alpha.Context.Proof.tree Tezos_protocol_environment_alpha.Context.Proof.t

module Make (Context : Sc_rollup_PVM_sig.Generic_pvm_context_sig) : S with type context = Context.Tree.t and type state = Context.tree and type proof = Context.proof