Module Tezos_raw_protocol_alpha.Sc_rollup_origination_machine

This module provides instantiation of all PVM kinds which can be used to perform rollup origination, but cannot be used to compute proofs or run computations.

module type S = sig ... end

module Arith : S

module Wasm : S

module Riscv : S