Module type Sc_rollup_origination_machine.S

include Sc_rollup_PVM_sig.PROTO_ORIGINATION

type state

The state of the PVM denotes a state of the rollup.

The life cycle of the PVM is as follows. It starts its execution from an initial_state. The initial state is specialized at origination with a boot_sector, using the install_boot_sector function. The resulting state is call the “genesis” of the rollup.

Afterwards, we classify states into two categories: "internal states" do not require any external information to be executed while "input states" are waiting for some information from the inbox to be executable.

type hash = Sc_rollup_repr.State_hash.t

A hash characterizes the contents of a state.

val state_hash : state -> hash Tezos_protocol_environment_alpha.Lwt.t

state_hash state returns a compressed representation of state.

val initial_state : empty:state -> state Tezos_protocol_environment_alpha.Lwt.t

initial_state ~empty is the initial state of the PVM, before its specialization with a given boot_sector. The initial state is built on the empty state which must be provided.

val install_boot_sector : state -> string -> state Tezos_protocol_environment_alpha.Lwt.t

install_boot_sector state boot_sector specializes the initial state of a PVM using a dedicated boot_sector, submitted at the origination of the rollup.

val empty_state : unit -> state