Module Sc_rollup_proof_repr.Dal_helpers

val import_level_is_valid : dal_activation_level:Raw_level_repr.t option -> dal_attestation_lag:int -> origination_level:Raw_level_repr.t -> commit_inbox_level:Raw_level_repr.t -> published_level:Raw_level_repr.t -> dal_attested_slots_validity_lag:int -> bool

We consider that a DAL page or slot published at a level published_level is valid, and produce or verify a proof for it if, and only if, the level is in the following boundaries:

DAL is activated and published_level >= dal_activation_level

published_level > origination_level: this means that the slot of the page was published after the rollup origination ;

published_level + dal_attestation_lag <= commit_inbox_level: this means that the slot of the page has been attested before or at the commit_inbox_level.

According to the definition in Sc_rollup_commitment_repr, commit_inbox_level (aka inbox_level in that module) is the level (included) up to which the PVM consumed all messages and DAL/DAC inputs before producing the related commitment. We also check that the given slot ID's index is within the range of allowed slots thanks to dal_number_of_slots.