Module Tezos_raw_protocol_alpha.Seed_storage

module Env = Tezos_protocol_environment_alpha.Environment

module Err = Env.Error_monad

type Err.error +=

| Unknown of {

oldest : Cycle_repr.t;

cycle : Cycle_repr.t;

latest : Cycle_repr.t;

}

val init : Raw_context.t -> Raw_context.t Err.tzresult Env.Lwt.t

Generates the first preserved_cycles+2 seeds for which there are no nonces.

val for_cycle : Raw_context.t -> Cycle_repr.t -> Seed_repr.seed Err.tzresult Env.Lwt.t

val cycle_end : Raw_context.t -> Cycle_repr.t -> (Raw_context.t * Nonce_storage.unrevealed list) Err.tzresult Env.Lwt.t

If it is the end of the cycle, computes and stores the seed of cycle at distance preserved_cycle+2 in the future using the seed of the previous cycle and the revelations of the current one.