Module Dal.Slot

module Headers : Storage_sigs.Non_iterable_indexed_data_storage with type t = Raw_context.t and type key = Raw_level_repr.t and type value = (Dal_slot_repr.Header.t * Contract_repr.t) list

This is a temporary storage for slot headers proposed onto the L1.

module History : Storage_sigs.Single_data_storage with type t := Raw_context.t and type value = Dal_slot_repr.History.t

This is a permanent storage for slot headers confirmed by the L1.

module LevelHistories : Storage_sigs.Single_data_storage with type t := Raw_context.t and type value = (Dal_slot_repr.History.Pointer_hash.t * Dal_slot_repr.History.t) list

This single entry stores the cells of the DAL skip list constructed during the block under validation. The list is expected to have exactly number_of_slots elements. Its cells ordering is not specified (and not relevant). A cell's binary encoding is bounded (the only part that is evolving in size over time is the number of backpointers, which is bounded by 64).