Module Tezos_raw_protocol_alpha.Storage_description

module Env = Tezos_protocol_environment_alpha.Environment

module Err = Env.Error_monad

type 'key t

Typed description of the key-value context.

val pp : Env.Format.formatter -> 'key t -> unit

Trivial display of the key-value context layout.

val build_directory : 'key t -> 'key Env.RPC_directory.t

Export an RPC hierarchy for querying the context. There is one service by possible path in the context. Services for "directory" are able to aggregate in one JSON object the whole subtree.

val create : unit -> 'key t

Create a empty context description, keys will be registered by side effects.

val register_value : 'key t -> get:('key -> 'a option Err.tzresult Env.Lwt.t) -> 'a Env.Data_encoding.t -> unit

Register a single key accessor at a given path.

val register_named_subcontext : 'key t -> string list -> 'key t

Return a description for a prefixed fragment of the given context. All keys registered in the subcontext will be shared by the external context

type (_, _, _) args =

| One : {

rpc_arg : 'a Env.RPC_arg.t;

encoding : 'a Env.Data_encoding.t;

compare : 'a -> 'a -> int;

} -> ('key, 'a, 'key * 'a) args

| Pair : ('key, 'a, 'inter_key) args * ('inter_key, 'b, 'sub_key) args -> ('key, 'a * 'b, 'sub_key) args

Description of an index as a sequence of `RPC_arg.t`.

val register_indexed_subcontext : 'key t -> list:('key -> 'arg list Err.tzresult Env.Lwt.t) -> ('key, 'arg, 'sub_key) args -> 'sub_key t

Return a description for a indexed sub-context. All keys registered in the subcontext will be shared by the external context. One should provide a function to list all the registered index in the context.

Helpers for manipulating and defining indexes.

val pack : ('key, 'a, 'sub_key) args -> 'key -> 'a -> 'sub_key

val unpack : ('key, 'a, 'sub_key) args -> 'sub_key -> 'key * 'a

module type INDEX = sig ... end