Module Storage_functors.Make_data_set_storage

Create storage for a compound type.

Parameters

module C : Raw_context.T

module I : INDEX

Signature

type t = C.t

type context = t

type elt = I.t

The type of elements.

val mem : context -> elt -> bool Tezos_protocol_environment_alpha.Lwt.t

Tells if an elt is a member of the set

val add : context -> elt -> Raw_context.t Tezos_protocol_environment_alpha.Lwt.t

Adds an elt as a member of the set

val remove : context -> elt -> Raw_context.t Tezos_protocol_environment_alpha.Lwt.t

Removes an elt from the set ; does nothing if not a member

val elements : context -> elt list Tezos_protocol_environment_alpha.Lwt.t

Returns the elements of the set, deserialized in a list in no particular order.

val fold : context -> order:[ `Sorted | `Undefined ] -> init:'a -> f:(elt -> 'a -> 'a Tezos_protocol_environment_alpha.Lwt.t) -> 'a Tezos_protocol_environment_alpha.Lwt.t

Iterates over the elements of the set.

val clear : context -> Raw_context.t Tezos_protocol_environment_alpha.Lwt.t

Removes all elements in the set