Module Unstake_requests_storage.For_RPC

Slow functions only used for RPCs

val apply_slash_to_unstaked_unfinalizable : Raw_context.t -> stored_requests -> (Cycle_repr.t * Tez_repr.t) list Tezos_protocol_environment_alpha.Error_monad.tzresult Tezos_protocol_environment_alpha.Lwt.t

Apply current slash history to unfinalizable unstake requests. prepare_finalize_unstake does not compute this value because it is never used internally. However, we need to apply slashes anyways when trying to compute the accurate balance of a staker