Module Tezos_raw_protocol_alpha.Voting_period_storage

module Env = Tezos_protocol_environment_alpha.Environment

module Err = Env.Error_monad

val init : Raw_context.t -> Voting_period_repr.t -> Raw_context.t Err.tzresult Env.Lwt.t

val init_first_period : Raw_context.t -> start_position:Env.Int32.t -> Raw_context.t Err.tzresult Env.Lwt.t

Sets the initial period to {voting_period = root; kind = Proposal; start_position}.

val reset : Raw_context.t -> Raw_context.t Err.tzresult Env.Lwt.t

Increment the index by one and set the kind to Proposal.

val succ : Raw_context.t -> Raw_context.t Err.tzresult Env.Lwt.t

Increment the index by one and set the kind to its successor.

val get_current : Raw_context.t -> Voting_period_repr.t Err.tzresult Env.Lwt.t

val get_current_kind : Raw_context.t -> Voting_period_repr.kind Err.tzresult Env.Lwt.t

val is_last_block : Raw_context.t -> bool Err.tzresult Env.Lwt.t

Returns true if the context level is the last of current voting period.

val get_rpc_current_info : Raw_context.t -> Voting_period_repr.info Err.tzresult Env.Lwt.t

Returns the voting period information for the current level.

val get_rpc_succ_info : Raw_context.t -> Voting_period_repr.info Err.tzresult Env.Lwt.t

Returns the voting period information for the next level.