Module Tezos_raw_protocol_alpha__Amendment

Only delegates with at least one roll take part in the amendment procedure. It works as follows:

Proposal period: delegates can submit protocol amendment proposals using the proposal operation. At the end of a proposal period, the proposal with most supporters is selected and we move to an exploration period. If there are no proposals, or a tie between proposals, a new proposal period starts.

Exploration period: delegates can cast votes to test or not the winning proposal using the ballot operation. At the end of an exploration period if participation reaches the quorum and the proposal has a supermajority in favor, we proceed to a cooldown period. Otherwise we go back to a proposal period. In any case, if there is enough participation the quorum is updated.

Cooldown period: Nothing happens, this period is only a time gap between exploration and promotion periods. At the end of a cooldown period we move to a promotion period.

Promotion period: delegates can cast votes to promote or not the proposal using the ballot operation. At the end of a promotion period if participation reaches the quorum and the proposal has a supermajority in favor, we move to an adoption period. Otherwise we go back to a proposal period. In any case, if there is enough participation the quorum is updated.

Adoption period: At the end of an adoption period, the proposal is activated as the new protocol.

val may_start_new_voting_period : Tezos_raw_protocol_alpha.Alpha_context.context -> Tezos_raw_protocol_alpha.Alpha_context.context Tezos_protocol_environment_alpha__Environment.Error_monad.tzresult Tezos_protocol_environment_alpha__Environment.Lwt.t

If at the end of a voting period, moves to the next one following the state machine of the amendment procedure.

type Tezos_protocol_environment_alpha__Environment.Error_monad.error += Unexpected_proposal | Unauthorized_proposal | Too_many_proposals | Empty_proposal

val record_proposals : Tezos_raw_protocol_alpha.Alpha_context.context -> Tezos_raw_protocol_alpha.Alpha_context.public_key_hash -> Tezos_protocol_environment_alpha__Environment.Protocol_hash.t list -> Tezos_raw_protocol_alpha.Alpha_context.context Tezos_protocol_environment_alpha__Environment.Error_monad.tzresult Tezos_protocol_environment_alpha__Environment.Lwt.t

Records a list of proposals for a delegate.

raises Unexpected_proposal

if ctxt is not in a proposal period.

raises Unauthorized_proposal

if delegate is not in the listing.

type Tezos_protocol_environment_alpha__Environment.Error_monad.error += Invalid_proposal | Unexpected_ballot | Unauthorized_ballot

val record_ballot : Tezos_raw_protocol_alpha.Alpha_context.context -> Tezos_raw_protocol_alpha.Alpha_context.public_key_hash -> Tezos_protocol_environment_alpha__Environment.Protocol_hash.t -> Tezos_raw_protocol_alpha.Alpha_context.Vote.ballot -> Tezos_raw_protocol_alpha.Alpha_context.context Tezos_protocol_environment_alpha__Environment.Error_monad.tzresult Tezos_protocol_environment_alpha__Environment.Lwt.t