Module Tezos_raw_protocol_alpha__Apply_results

Result of applying an operation, can be used for experimenting with protocol updates, by clients to print out a summary of the operation at pre-injection simulation and at confirmation time, and by block explorers.

type 'kind operation_metadata = {

contents : 'kind contents_result_list;

}

Result of applying a Operation.t. Follows the same structure.

and packed_operation_metadata =

| Operation_metadata : 'kind operation_metadata -> packed_operation_metadata

| No_operation_metadata : packed_operation_metadata

and 'kind contents_result_list =

| Single_result : 'kind contents_result -> 'kind contents_result_list

| Cons_result : 'kind Tezos_raw_protocol_alpha.Alpha_context.Kind.manager contents_result * 'rest Tezos_raw_protocol_alpha.Alpha_context.Kind.manager contents_result_list -> ('kind * 'rest) Tezos_raw_protocol_alpha.Alpha_context.Kind.manager contents_result_list

Result of applying a Operation.contents_list. Follows the same structure.

and packed_contents_result_list =

| Contents_result_list : 'kind contents_result_list -> packed_contents_result_list

and 'kind contents_result =

| Endorsement_result : {

balance_updates : Tezos_raw_protocol_alpha.Alpha_context.Receipt.balance_updates;

delegate : Tezos_protocol_environment_alpha__Environment.Signature.Public_key_hash.t;

slots : int list;

} -> Tezos_raw_protocol_alpha.Alpha_context.Kind.endorsement contents_result

| Seed_nonce_revelation_result : Tezos_raw_protocol_alpha.Alpha_context.Receipt.balance_updates -> Tezos_raw_protocol_alpha.Alpha_context.Kind.seed_nonce_revelation contents_result

| Endorsement_with_slot_result : Tezos_raw_protocol_alpha.Alpha_context.Kind.endorsement contents_result -> Tezos_raw_protocol_alpha.Alpha_context.Kind.endorsement_with_slot contents_result

| Double_endorsement_evidence_result : Tezos_raw_protocol_alpha.Alpha_context.Receipt.balance_updates -> Tezos_raw_protocol_alpha.Alpha_context.Kind.double_endorsement_evidence contents_result

| Double_baking_evidence_result : Tezos_raw_protocol_alpha.Alpha_context.Receipt.balance_updates -> Tezos_raw_protocol_alpha.Alpha_context.Kind.double_baking_evidence contents_result

| Activate_account_result : Tezos_raw_protocol_alpha.Alpha_context.Receipt.balance_updates -> Tezos_raw_protocol_alpha.Alpha_context.Kind.activate_account contents_result

| Proposals_result : Tezos_raw_protocol_alpha.Alpha_context.Kind.proposals contents_result

| Ballot_result : Tezos_raw_protocol_alpha.Alpha_context.Kind.ballot contents_result

| Manager_operation_result : {

balance_updates : Tezos_raw_protocol_alpha.Alpha_context.Receipt.balance_updates;

operation_result : 'kind manager_operation_result;

internal_operation_results : packed_internal_operation_result list;

} -> 'kind Tezos_raw_protocol_alpha.Alpha_context.Kind.manager contents_result

Result of applying an Operation.contents. Follows the same structure.

and packed_contents_result =

| Contents_result : 'kind contents_result -> packed_contents_result

and 'kind manager_operation_result =

| Applied of 'kind successful_manager_operation_result

| Backtracked of 'kind successful_manager_operation_result * Tezos_protocol_environment_alpha__Environment.Error_monad.error Tezos_protocol_environment_alpha__Environment.Error_monad.trace option

| Failed : 'kind Tezos_raw_protocol_alpha.Alpha_context.Kind.manager * Tezos_protocol_environment_alpha__Environment.Error_monad.error Tezos_protocol_environment_alpha__Environment.Error_monad.trace -> 'kind manager_operation_result

| Skipped : 'kind Tezos_raw_protocol_alpha.Alpha_context.Kind.manager -> 'kind manager_operation_result

The result of an operation in the queue. Skipped ones should always be at the tail, and after a single Failed.

and _ successful_manager_operation_result =

| Reveal_result : {

consumed_gas : Tezos_raw_protocol_alpha.Alpha_context.Gas.Arith.fp;

} -> Tezos_raw_protocol_alpha.Alpha_context.Kind.reveal successful_manager_operation_result

| Transaction_result : {

storage : Tezos_raw_protocol_alpha.Alpha_context.Script.expr option;

lazy_storage_diff : Tezos_raw_protocol_alpha.Alpha_context.Lazy_storage.diffs option;

balance_updates : Tezos_raw_protocol_alpha.Alpha_context.Receipt.balance_updates;

originated_contracts : Tezos_raw_protocol_alpha.Alpha_context.Contract.t list;

consumed_gas : Tezos_raw_protocol_alpha.Alpha_context.Gas.Arith.fp;

storage_size : Tezos_protocol_environment_alpha__Environment.Z.t;

paid_storage_size_diff : Tezos_protocol_environment_alpha__Environment.Z.t;

allocated_destination_contract : bool;

} -> Tezos_raw_protocol_alpha.Alpha_context.Kind.transaction successful_manager_operation_result

| Origination_result : {

lazy_storage_diff : Tezos_raw_protocol_alpha.Alpha_context.Lazy_storage.diffs option;

balance_updates : Tezos_raw_protocol_alpha.Alpha_context.Receipt.balance_updates;

originated_contracts : Tezos_raw_protocol_alpha.Alpha_context.Contract.t list;

consumed_gas : Tezos_raw_protocol_alpha.Alpha_context.Gas.Arith.fp;

storage_size : Tezos_protocol_environment_alpha__Environment.Z.t;

paid_storage_size_diff : Tezos_protocol_environment_alpha__Environment.Z.t;

} -> Tezos_raw_protocol_alpha.Alpha_context.Kind.origination successful_manager_operation_result

| Delegation_result : {

consumed_gas : Tezos_raw_protocol_alpha.Alpha_context.Gas.Arith.fp;

} -> Tezos_raw_protocol_alpha.Alpha_context.Kind.delegation successful_manager_operation_result

Result of applying a manager_operation_content, either internal or external.

and packed_successful_manager_operation_result =

| Successful_manager_result : 'kind successful_manager_operation_result -> packed_successful_manager_operation_result

and packed_internal_operation_result =

| Internal_operation_result : 'kind Tezos_raw_protocol_alpha.Alpha_context.internal_operation * 'kind manager_operation_result -> packed_internal_operation_result

val pack_migration_operation_results : Tezos_raw_protocol_alpha.Alpha_context.Migration.origination_result list -> packed_successful_manager_operation_result list

val operation_metadata_encoding : packed_operation_metadata Tezos_protocol_environment_alpha__Environment.Data_encoding.t

Serializer for packed_operation_result.

val operation_data_and_metadata_encoding : (Tezos_raw_protocol_alpha.Alpha_context.Operation.packed_protocol_data * packed_operation_metadata) Tezos_protocol_environment_alpha__Environment.Data_encoding.t

type 'kind contents_and_result_list =

| Single_and_result : 'kind Tezos_raw_protocol_alpha.Alpha_context.contents * 'kind contents_result -> 'kind contents_and_result_list

| Cons_and_result : 'kind Tezos_raw_protocol_alpha.Alpha_context.Kind.manager Tezos_raw_protocol_alpha.Alpha_context.contents * 'kind Tezos_raw_protocol_alpha.Alpha_context.Kind.manager contents_result * 'rest Tezos_raw_protocol_alpha.Alpha_context.Kind.manager contents_and_result_list -> ('kind * 'rest) Tezos_raw_protocol_alpha.Alpha_context.Kind.manager contents_and_result_list

type packed_contents_and_result_list =

| Contents_and_result_list : 'kind contents_and_result_list -> packed_contents_and_result_list

val contents_and_result_list_encoding : packed_contents_and_result_list Tezos_protocol_environment_alpha__Environment.Data_encoding.t

val pack_contents_list : 'kind Tezos_raw_protocol_alpha.Alpha_context.contents_list -> 'kind contents_result_list -> 'kind contents_and_result_list

val unpack_contents_list : 'kind contents_and_result_list -> 'kind Tezos_raw_protocol_alpha.Alpha_context.contents_list * 'kind contents_result_list

val to_list : packed_contents_result_list -> packed_contents_result list

val of_list : packed_contents_result list -> packed_contents_result_list

type ('a, 'b) eq =

| Eq : ('a, 'a) eq

val kind_equal_list : 'kind Tezos_raw_protocol_alpha.Alpha_context.contents_list -> 'kind2 contents_result_list -> ('kind, 'kind2) eq option

type block_metadata = {

baker : Tezos_protocol_environment_alpha__Environment.Signature.Public_key_hash.t;

level_info : Tezos_raw_protocol_alpha.Alpha_context.Level.t;

voting_period_info : Tezos_raw_protocol_alpha.Alpha_context.Voting_period.info;

nonce_hash : Tezos_raw_protocol_alpha.Nonce_hash.t option;

consumed_gas : Tezos_raw_protocol_alpha.Alpha_context.Gas.Arith.fp;

deactivated : Tezos_protocol_environment_alpha__Environment.Signature.Public_key_hash.t list;

balance_updates : Tezos_raw_protocol_alpha.Alpha_context.Receipt.balance_updates;

liquidity_baking_escape_ema : Tezos_raw_protocol_alpha.Alpha_context.Liquidity_baking.escape_ema;

implicit_operations_results : packed_successful_manager_operation_result list;

}

val block_metadata_encoding : block_metadata Tezos_protocol_environment_alpha__Environment.Data_encoding.encoding