Module Tezos_raw_protocol_alpha__Level_storage

val current : Tezos_raw_protocol_alpha.Raw_context.t -> Tezos_raw_protocol_alpha.Level_repr.t

val previous : Tezos_raw_protocol_alpha.Raw_context.t -> Tezos_raw_protocol_alpha.Level_repr.t

val root : Tezos_raw_protocol_alpha.Raw_context.t -> Tezos_raw_protocol_alpha.Level_repr.t

val from_raw : Tezos_raw_protocol_alpha.Raw_context.t -> ?⁠offset:int32 -> Tezos_raw_protocol_alpha.Raw_level_repr.t -> Tezos_raw_protocol_alpha.Level_repr.t

val pred : Tezos_raw_protocol_alpha.Raw_context.t -> Tezos_raw_protocol_alpha.Level_repr.t -> Tezos_raw_protocol_alpha.Level_repr.t option

val succ : Tezos_raw_protocol_alpha.Raw_context.t -> Tezos_raw_protocol_alpha.Level_repr.t -> Tezos_raw_protocol_alpha.Level_repr.t

val first_level_in_cycle : Tezos_raw_protocol_alpha.Raw_context.t -> Tezos_raw_protocol_alpha.Cycle_repr.t -> Tezos_raw_protocol_alpha.Level_repr.t

val last_level_in_cycle : Tezos_raw_protocol_alpha.Raw_context.t -> Tezos_raw_protocol_alpha.Cycle_repr.t -> Tezos_raw_protocol_alpha.Level_repr.t

val levels_in_cycle : Tezos_raw_protocol_alpha.Raw_context.t -> Tezos_raw_protocol_alpha.Cycle_repr.t -> Tezos_raw_protocol_alpha.Level_repr.t list

val levels_in_current_cycle : Tezos_raw_protocol_alpha.Raw_context.t -> ?⁠offset:int32 -> unit -> Tezos_raw_protocol_alpha.Level_repr.t list

val levels_with_commitments_in_cycle : Tezos_raw_protocol_alpha.Raw_context.t -> Tezos_raw_protocol_alpha.Cycle_repr.t -> Tezos_raw_protocol_alpha.Level_repr.t list

val last_allowed_fork_level : Tezos_raw_protocol_alpha.Raw_context.t -> Tezos_raw_protocol_alpha.Raw_level_repr.t

val dawn_of_a_new_cycle : Tezos_raw_protocol_alpha.Raw_context.t -> Tezos_raw_protocol_alpha.Cycle_repr.t option

Returns Some cycle if the current level represents the last level of cycle and None if the level is not the last level of a cycle.

val may_snapshot_rolls : Tezos_raw_protocol_alpha.Raw_context.t -> bool

Returns true if the rolls should be snapshot at the current level.