Module Tezos_raw_protocol_alpha__Script_interpreter

This is the Michelson interpreter.

This module offers a way to execute either a Michelson script or a Michelson instruction.

Implementation details are documented in the .ml file.

type Tezos_protocol_environment_alpha__Environment.Error_monad.error += Reject of Tezos_raw_protocol_alpha.Alpha_context.Script.location * Tezos_raw_protocol_alpha.Alpha_context.Script.expr * Tezos_raw_protocol_alpha.Script_typed_ir.execution_trace option

type Tezos_protocol_environment_alpha__Environment.Error_monad.error += Overflow of Tezos_raw_protocol_alpha.Alpha_context.Script.location * Tezos_raw_protocol_alpha.Script_typed_ir.execution_trace option

type Tezos_protocol_environment_alpha__Environment.Error_monad.error += Runtime_contract_error : Tezos_raw_protocol_alpha.Alpha_context.Contract.t * Tezos_raw_protocol_alpha.Alpha_context.Script.expr -> Tezos_protocol_environment_alpha__Environment.Error_monad.error

type Tezos_protocol_environment_alpha__Environment.Error_monad.error += Bad_contract_parameter of Tezos_raw_protocol_alpha.Alpha_context.Contract.t

type Tezos_protocol_environment_alpha__Environment.Error_monad.error += Cannot_serialize_failure

type Tezos_protocol_environment_alpha__Environment.Error_monad.error += Cannot_serialize_storage

type Tezos_protocol_environment_alpha__Environment.Error_monad.error += Michelson_too_many_recursive_calls

type execution_result = {

ctxt : Tezos_raw_protocol_alpha.Alpha_context.context;

storage : Tezos_raw_protocol_alpha.Alpha_context.Script.expr;

lazy_storage_diff : Tezos_raw_protocol_alpha.Alpha_context.Lazy_storage.diffs option;

operations : Tezos_raw_protocol_alpha.Alpha_context.packed_internal_operation list;

}

type step_constants = {

source : Tezos_raw_protocol_alpha.Alpha_context.Contract.t;

payer : Tezos_raw_protocol_alpha.Alpha_context.Contract.t;

self : Tezos_raw_protocol_alpha.Alpha_context.Contract.t;

amount : Tezos_raw_protocol_alpha.Alpha_context.Tez.t;

chain_id : Tezos_protocol_environment_alpha__Environment.Chain_id.t;

}

val step : Tezos_raw_protocol_alpha.Script_typed_ir.logger option -> Tezos_raw_protocol_alpha.Alpha_context.context -> step_constants -> ('a, 's, 'r, 'f) Tezos_raw_protocol_alpha.Script_typed_ir.kdescr -> 'a -> 's -> ('r * 'f * Tezos_raw_protocol_alpha.Alpha_context.context) Tezos_protocol_environment_alpha__Environment.Error_monad.tzresult Tezos_protocol_environment_alpha__Environment.Lwt.t

val execute : ?⁠logger:Tezos_raw_protocol_alpha.Script_typed_ir.logger -> Tezos_raw_protocol_alpha.Alpha_context.t -> Tezos_raw_protocol_alpha.Script_ir_translator.unparsing_mode -> step_constants -> script:Tezos_raw_protocol_alpha.Alpha_context.Script.t -> entrypoint:string -> parameter:Tezos_raw_protocol_alpha.Alpha_context.Script.expr -> internal:bool -> execution_result Tezos_protocol_environment_alpha__Environment.Error_monad.tzresult Tezos_protocol_environment_alpha__Environment.Lwt.t

val kstep : Tezos_raw_protocol_alpha.Script_typed_ir.logger option -> Tezos_raw_protocol_alpha.Alpha_context.context -> step_constants -> ('a, 's, 'r, 'f) Tezos_raw_protocol_alpha.Script_typed_ir.kinstr -> 'a -> 's -> ('r * 'f * Tezos_raw_protocol_alpha.Alpha_context.context) Tezos_protocol_environment_alpha__Environment.Error_monad.tzresult Tezos_protocol_environment_alpha__Environment.Lwt.t

kstep logger ctxt step_constants kinstr accu stack interprets the script represented by kinstr under the context ctxt. This will turn a stack whose topmost element is accu and remaining elements stack into a new accumulator and a new stack. This function also returns an updated context. If logger is given, kstep calls back its functions at specific points of the execution. The execution is parameterized by some step_constants.

module Internals : sig ... end